Unified Setup/TearDown logic into "Setup.cs"
