+++ test/Libraries/CoreNodesTests/Setup.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.IO;
+using System.Reflection;
 using Dynamo;
 using Dynamo.Utilities;
 using NUnit.Framework;
@@ -8,15 +10,29 @@ namespace DSCoreNodesTests
     [SetUpFixture]
     public class Setup
     {
+        private AssemblyHelper assemblyHelper;
+
         [SetUp]
-        public void RunBeforeAnyTests()
+        public void SetUp()
         {
-            AppDomain.CurrentDomain.AssemblyResolve += AssemblyHelper.ResolveAssembly;
+            var assemblyPath = Assembly.GetExecutingAssembly().Location;
+            var moduleRootFolder = Path.GetDirectoryName(assemblyPath);
+
+            var resolutionPaths = new[]
+            {
+                // These tests need "DSCoreNodesUI.dll" under "nodes" folder.
+                Path.Combine(moduleRootFolder, "nodes")
+            };
+
+            assemblyHelper = new AssemblyHelper(moduleRootFolder, resolutionPaths);
+            AppDomain.CurrentDomain.AssemblyResolve += assemblyHelper.ResolveAssemblyNew;
         }
 
         [TearDown]
-        public void RunAfterAnyTests()
+        public void TearDown()
         {
+            AppDomain.CurrentDomain.AssemblyResolve -= assemblyHelper.ResolveAssemblyNew;
+            assemblyHelper = null;
         }
     }
 }
